+++ db_vlvm/db_image_homography.h
@@ -43,9 +43,9 @@ although desirable for numerical conditioning
 \param xp3  image 2 point 3
 \param xp4  image 2 point 4
 */
-DB_API void db_StitchProjective2D_4Points(double H[9],
-                                      double x1[3],double x2[3],double x3[3],double x4[3],
-                                      double xp1[3],double xp2[3],double xp3[3],double xp4[3]);
+DB_API void db_StitchProjective2D_4Points(float H[9],
+                                      float x1[3],float x2[3],float x3[3],float x4[3],
+                                      float xp1[3],float xp2[3],float xp3[3],float xp4[3]);
 
 /*!
 Solve for affine H such that xp~Hx. Prior normalization is not necessary,
@@ -58,9 +58,9 @@ although desirable for numerical conditioning
 \param xp2  image 2 point 2
 \param xp3  image 2 point 3
 */
-DB_API void db_StitchAffine2D_3Points(double H[9],
-                                      double x1[3],double x2[3],double x3[3],
-                                      double xp1[3],double xp2[3],double xp3[3]);
+DB_API void db_StitchAffine2D_3Points(float H[9],
+                                      float x1[3],float x2[3],float x3[3],
+                                      float xp1[3],float xp2[3],float xp3[3]);
 
 /*!
 Solve for rotation R such that xp~Rx.
@@ -71,15 +71,15 @@ Image points have to be of unit norm for the least squares to be meaningful.
 \param xp1  image 2 point 1
 \param xp2  image 2 point 2
 */
-inline void db_StitchCameraRotation_2Points(double R[9],
+inline void db_StitchCameraRotation_2Points(float R[9],
                                             /*Image points have to be of unit norm
                                             for the least squares to be meaningful*/
-                                            double x1[3],double x2[3],
-                                            double xp1[3],double xp2[3])
+                                            float x1[3],float x2[3],
+                                            float xp1[3],float xp2[3])
 {
-    double* x[2];
-    double* xp[2];
-    double scale,t[3];
+    float* x[2];
+    float* xp[2];
+    float scale,t[3];
 
     x[0]=x1;
     x[1]=x2;
@@ -97,8 +97,8 @@ is required, although it could be desirable to keep x1,x2,xp1 and xp2 of reasona
 If a solution is obtained the function returns 1, otherwise 0. If the focal length is desired
 a valid pointer should be passed in f
 */
-DB_API int db_StitchRotationCommonFocalLength_3Points(double H[9],double x1[3],double x2[3],double x3[3],
-                                                      double xp1[3],double xp2[3],double xp3[3],double *f=0,int signed_disambiguation=1);
+DB_API int db_StitchRotationCommonFocalLength_3Points(float H[9],float x1[3],float x2[3],float x3[3],
+                                                      float xp1[3],float xp2[3],float xp3[3],float *f=0,int signed_disambiguation=1);
 
 /*!
 Find scale, rotation and translation of the similarity that
@@ -152,8 +152,8 @@ With orientation_preserving:
 \param allow_rotation   compute rotation (if 0, R=[I])
 \param allow_translation compute translation (if 0 t = [0,0]')
 */
-DB_API void db_StitchSimilarity2DRaw(double *scale,double R[4],double t[2],
-                            double **Xp,double **X,int nr_points,int orientation_preserving=1,
+DB_API void db_StitchSimilarity2DRaw(float *scale,float R[4],float t[2],
+                            float **Xp,float **X,int nr_points,int orientation_preserving=1,
                             int allow_scaling=1,int allow_rotation=1,int allow_translation=1);
 /*!
 See db_StitchRotationCommonFocalLength_3Points().
@@ -166,10 +166,10 @@ See db_StitchRotationCommonFocalLength_3Points().
 \param allow_rotation   compute rotation (if 0, R=[I])
 \param allow_translation compute translation (if 0 t = [0,0]')
 */
-inline void db_StitchSimilarity2D(double H[9],double **Xp,double **X,int nr_points,int orientation_preserving=1,
+inline void db_StitchSimilarity2D(float H[9],float **Xp,float **X,int nr_points,int orientation_preserving=1,
                                   int allow_scaling=1,int allow_rotation=1,int allow_translation=1)
 {
-    double s,R[4],t[2];
+    float s,R[4],t[2];
 
     db_StitchSimilarity2DRaw(&s,R,t,Xp,X,nr_points,orientation_preserving,
         allow_scaling,allow_rotation,allow_translation);
